Merge Pull-Request from Upstream (duplicate confirmation mail) (#9)
* Update README.md

* Fix Prevent duplicate confirmation emails to customers

If the asynchronous sending of emails is disabled,  in most cases Magento 2.3 will send two confirmation emails to the customer. The additional query of getEmailSent should prevent this.
The error does not occur with asynchronous mail sending. However, due to the catastrophic implementation  of the new inventory in Magento 2.3, this is no longer an option.